Drop core router module
