+++ bloomerp/django_bloomerp/bloomerp/views/core/base_detail.py
@@ -1,0 +1,16 @@
+from typing import Any
+from django.views.generic.detail import DetailView
+from bloomerp.views.mixins import BloomerpModelContextMixin, HtmxMixin
+
+
+class BloomerpBaseDetailView(HtmxMixin, BloomerpModelContextMixin, DetailView):
+    htmx_template = "bloomerp_htmx_base_view.html"
+    tabs = None
+    exclude_header = False
+
+    def get_context_data(self, **kwargs: Any) -> dict:
+        context = super().get_context_data(**kwargs)
+        context["exclude_header"] = self.exclude_header
+        if self.tabs:
+            context["tabs"] = self.tabs
+        return context
